refactor: 重新定義API型別
