+++ src/apis/responseTypes.d.mts
@@ -1,0 +1,9 @@
+export interface googleTokenType {
+    data: {
+        access_token: string;
+        expires_in: number;
+        refresh_token: string;
+        scope: string;
+        token_type: string;
+    };
+}
